backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
twilio>=9.0.0
python-dotenv>=1.0.1
gspread>=6.0.0
google-auth>=2.28.0
